+++ ErsatzTV.Core/Interfaces/Scheduling/IYamlScheduleValidator.cs
@@ -1,0 +1,6 @@
+namespace ErsatzTV.Core.Interfaces.Scheduling;
+
+public interface IYamlScheduleValidator
+{
+    Task<bool> ValidateSchedule(string yaml);
+}