+++ ErsatzTV.Core/Scheduling/YamlScheduling/YamlPlayoutBuilder.cs
@@ -7,6 +7,7 @@ using ErsatzTV.Core.Interfaces.Scheduling;
 using ErsatzTV.Core.Scheduling.YamlScheduling.Handlers;
 using ErsatzTV.Core.Scheduling.YamlScheduling.Models;
 using ErsatzTV.Core.Search;
+using LanguageExt.UnsafeValueAccess;
 using Microsoft.Extensions.Logging;
 using YamlDotNet.Serialization;
 using YamlDotNet.Serialization.NamingConventions;
@@ -18,6 +19,7 @@ public class YamlPlayoutBuilder(
     IConfigElementRepository configElementRepository,
     IMediaCollectionRepository mediaCollectionRepository,
     IChannelRepository channelRepository,
+    IYamlScheduleValidator yamlScheduleValidator,
     ILogger<YamlPlayoutBuilder> logger)
     : IYamlPlayoutBuilder
 {
@@ -29,7 +31,15 @@ public class YamlPlayoutBuilder(
             return playout;
         }
 
-        YamlPlayoutDefinition playoutDefinition = await LoadYamlDefinition(playout, cancellationToken);
+        Option<YamlPlayoutDefinition> maybePlayoutDefinition = await LoadYamlDefinition(playout, cancellationToken);
+        if (maybePlayoutDefinition.IsNone)
+        {
+            logger.LogWarning("YAML playout file {File} is invalid; aborting.", playout.TemplateFile);
+            return playout;
+        }
+
+        // using ValueUnsafe to avoid nesting
+        YamlPlayoutDefinition playoutDefinition = maybePlayoutDefinition.ValueUnsafe();
 
         DateTimeOffset start = DateTimeOffset.Now;
 
@@ -62,12 +72,7 @@ public class YamlPlayoutBuilder(
         {
             foreach (PlayoutAnchor prevAnchor in Optional(playout.Anchor))
             {
-                // TODO: does this matter?
-                //context.GuideGroup = prevAnchor.NextGuideGroup;
-
-                context.CurrentTime = new DateTimeOffset(prevAnchor.NextStart.ToLocalTime(), start.Offset);
-
-                context.InstructionIndex = prevAnchor.NextInstructionIndex;
+                context.Reset(prevAnchor, start);
             }
         }
         else
@@ -125,7 +130,13 @@ public class YamlPlayoutBuilder(
                     }
                     else
                     {
-                        await handler.Handle(context, instruction, mode, logger, cancellationToken);
+                        await handler.Handle(
+                            context,
+                            instruction,
+                            mode,
+                            _ => Task.CompletedTask,
+                            logger,
+                            cancellationToken);
                     }
                 }
             }
@@ -167,10 +178,20 @@ public class YamlPlayoutBuilder(
 
             foreach (IYamlPlayoutHandler handler in maybeHandler)
             {
-                if (!await handler.Handle(context, instruction, mode, logger, cancellationToken))
+                if (!await handler.Handle(context, instruction, mode, ExecuteSequenceLocal, logger, cancellationToken))
                 {
                     logger.LogInformation("YAML playout instruction handler failed");
                 }
+
+                continue;
+
+                async Task ExecuteSequenceLocal(string sequence) => await ExecuteSequence(
+                    handlers,
+                    enumeratorCache,
+                    mode,
+                    context,
+                    sequence,
+                    cancellationToken);
             }
 
             if (!instruction.ChangesIndex)
@@ -191,8 +212,7 @@ public class YamlPlayoutBuilder(
         var anchor = new PlayoutAnchor
         {
             NextStart = maxTime,
-            NextInstructionIndex = context.InstructionIndex,
-            NextGuideGroup = context.PeekNextGuideGroup()
+            Context = context.Serialize()
         };
 
         context.AdvanceGuideGroup();
@@ -205,6 +225,37 @@ public class YamlPlayoutBuilder(
         playout.Anchor = anchor;
 
         return playout;
+    }
+
+    private async Task ExecuteSequence(
+        Dictionary<YamlPlayoutInstruction, IYamlPlayoutHandler> handlers,
+        EnumeratorCache enumeratorCache,
+        PlayoutBuildMode mode,
+        YamlPlayoutContext context,
+        string sequence,
+        CancellationToken cancellationToken)
+    {
+        var sequenceInstructions = context.Definition.Sequence
+            .Filter(s => s.Key == sequence)
+            .HeadOrNone()
+            .Map(s => s.Items)
+            .Flatten()
+            .ToList();
+
+        foreach (YamlPlayoutInstruction instruction in sequenceInstructions)
+        {
+            //logger.LogDebug("Current playout instruction: {Instruction}", instruction.GetType().Name);
+
+            Option<IYamlPlayoutHandler> maybeHandler = GetHandlerForInstruction(handlers, enumeratorCache, instruction);
+
+            foreach (IYamlPlayoutHandler handler in maybeHandler)
+            {
+                if (!await handler.Handle(context, instruction, mode, _ => Task.CompletedTask, logger, cancellationToken))
+                {
+                    logger.LogInformation("YAML playout instruction handler failed");
+                }
+            }
+        }
     }
 
     private static bool DetectCycle(YamlPlayoutDefinition definition)
@@ -292,6 +343,7 @@ public class YamlPlayoutBuilder(
             YamlPlayoutEpgGroupInstruction => new YamlPlayoutEpgGroupHandler(),
             YamlPlayoutWatermarkInstruction => new YamlPlayoutWatermarkHandler(channelRepository),
             YamlPlayoutShuffleSequenceInstruction => new YamlPlayoutShuffleSequenceHandler(),
+            YamlPreRollInstruction => new YamlPlayoutPreRollHandler(),
 
             YamlPlayoutSkipItemsInstruction => new YamlPlayoutSkipItemsHandler(enumeratorCache),
             YamlPlayoutSkipToItemInstruction => new YamlPlayoutSkipToItemHandler(enumeratorCache),
@@ -314,11 +366,15 @@ public class YamlPlayoutBuilder(
         return Optional(handler);
     }
 
-    private async Task<YamlPlayoutDefinition> LoadYamlDefinition(Playout playout, CancellationToken cancellationToken)
+    private async Task<Option<YamlPlayoutDefinition>> LoadYamlDefinition(Playout playout, CancellationToken cancellationToken)
     {
         try
         {
             string yaml = await File.ReadAllTextAsync(playout.TemplateFile, cancellationToken);
+            if (await yamlScheduleValidator.ValidateSchedule(yaml) == false)
+            {
+                return Option<YamlPlayoutDefinition>.None;
+            }
 
             IDeserializer deserializer = new DeserializerBuilder()
                 .WithNamingConvention(CamelCaseNamingConvention.Instance)
@@ -346,6 +402,7 @@ public class YamlPlayoutBuilder(
                         { "watermark", typeof(YamlPlayoutWatermarkInstruction) },
                         { "pad_to_next", typeof(YamlPlayoutPadToNextInstruction) },
                         { "pad_until", typeof(YamlPlayoutPadUntilInstruction) },
+                        { "pre_roll", typeof(YamlPreRollInstruction) },
                         { "repeat", typeof(YamlPlayoutRepeatInstruction) },
                         { "sequence", typeof(YamlPlayoutSequenceInstruction) },
                         { "shuffle_sequence", typeof(YamlPlayoutShuffleSequenceInstruction) },