+++ ErsatzTV/Startup.cs
@@ -730,6 +730,7 @@ public class Startup
         services.AddScoped<IJellyfinSecretStore, JellyfinSecretStore>();
         services.AddScoped<IEmbySecretStore, EmbySecretStore>();
         services.AddScoped<IScriptEngine, ScriptEngine>();
+        services.AddScoped<IYamlScheduleValidator, YamlScheduleValidator>();
 
         services.AddScoped<PlexEtag>();